solarwinds-mcp-server/pyproject.toml
[project]
name = "solarwinds-mcp"
version = "0.1.0"
description = "Read-only MCP server for SolarWinds Service Desk"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
solarwinds-mcp = "solarwinds_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/solarwinds_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.22"]
